[hotfix][table] Refactor DataGen source to use LogicalTypeVisitor pattern
